♻️  intent_detection.py 추출 — 13 intent matchers + phrase tuples (P0-L step 6)
변경 이유
- `detect_engineering_intent` 의 분기 우선순위가 P0-J / P0-K 에서 자주
  손댄 영역이라, intent 추가 / 우선순위 조정 시 변경 표면이 한 파일에
  몰려 있어야 한다. 가장 큰 책임 단위를 단독 모듈로 분리.

주요 변경 사항
- 새 파일 `engineering_conversation/intent_detection.py` (약 570 줄):
  - `detect_engineering_intent` 메인 분기
  - `split_task_branches`, `_normalize`
  - 13 phrase tuple (`_CONFIRMATION_PHRASES`/`_STANDALONE`,
    `_STATUS_DIAGNOSTIC_PHRASES`, `_SESSION_COUNT_PHRASES`,
    `_SESSION_LIST_PHRASES`, `_BLOCKED_REASON_PHRASES`,
    `_CONTINUE_EXISTING_PHRASES`, `_CHANGE_DIRECTION_PHRASES`,
    `_GENERAL_HELP_PHRASES`, `_VAGUE_TOKEN_RUNS`, `_SPLIT_PATTERN`)
  - 11 matcher helper (`_is_*`, `_asks_for_general_help`,
    `_looks_too_vague`, `_asks_to_*_thread`)
- `_legacy.py` 의 detection 블록 (약 485 줄) 제거 후 re-export 만 유지.
- `task_shaping.py` / `status_responses.py` 의 lazy `from ._legacy import`
  를 `.intent_detection` 직접 import 로 교체 — `_legacy` 의존 끊김.
- `__init__.py` facade 가 `detect_engineering_intent` /
  `split_task_branches` 도 `.intent_detection` 직접 import 으로 끌어옴.

비고
- 4657 PASS + 1 skipped 회귀 유지 (동작 변경 0).
- 다음 step (7): research_bootstrap.py 추출.
